+++ tests/test_driver.py
@@ -51,3 +51,45 @@ def test_driver(squashfs_file):
     # Reading subdir/file2.txt
     with fs.open("subdir/file2.txt", "rb") as f:
         assert f.read().decode() == "Hello from file 2 in subdir"
+
+
+def test_close_semantics(squashfs_file):
+    fs = SquashFSFileSystem(squashfs_file)
+    assert not fs.closed
+
+    f = fs.open("file1.txt", "rb")
+    assert not f.closed
+    assert not fs.fo.closed
+
+    f.close()
+    assert f.closed
+    assert not fs.fo.closed
+    with pytest.raises(ValueError, match="I/O operation on closed file\\."):
+        f.read(1)
+
+    with fs.open("file1.txt", "rb") as g:
+        assert not g.closed
+        _ = g.read(1)
+
+    assert g.closed
+    assert not fs.fo.closed
+
+    fs.close()
+    assert fs.closed
+    assert fs.fo.closed
+    with pytest.raises(
+        ValueError, match="I/O operation on closed filesystem\\."
+    ):
+        fs.open("file1.txt", "rb")
+
+
+def test_nested_context_managers(squashfs_file):
+    with SquashFSFileSystem(squashfs_file) as fs:
+        assert not fs.closed
+        assert not fs.fo.closed
+        with fs.open("file1.txt", "rb") as f:
+            assert f.read().decode() == "Hello from file 1"
+        assert f.closed
+
+    assert fs.closed
+    assert fs.fo.closed